requirements.txt
fastapi>=0.109.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
pyyaml>=6.0.1
pillow>=10.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=8.0.0